CharacterSpriteBehavior: Handle left- and front-facing sprites
Previously, this behavior assumed that the animations it controls face right.

The majority of sprites used in the game do face right, but there are
some exceptions: for example, the musician faces forward, and the
knitwitch faces left.

Of course, neither of these sprites have a walk animation; and if we did
want the KnitWitch to move, we could invert the asset. But you could
imagine having a sprite that can move but should not be flipped based on
its direction (perhaps a lightning storm).

Add a new property to CharacterSpriteBehavior, indicating which way the
sprite faces. Use this property to not flip the sprite if it faces
UNSPECIFIED; and invert the current flip_h logic if the resting
direction is LEFT rather than RIGHT. Set the default to RIGHT, matching
the convention for characters in the game which can move.