AreaFiller: Integrate with editor undo/redo
Previously, it was not possible to undo the "Refill" action, other than
with `git restore` or similar.

Use EditorUndoRedoManager to add the refill action to the editor's
undo/redo stack. This is intended to be used from EditorPlugins, which
this script is not, but since the EditorUndoRedoManager is actually a
singleton we can create a temporary EditorPlugin instance, get the
EditorUndoRedoManager from it, and free it.

Because the EditorPlugin and EditorUndoRedoManager classes are only
available in the editor, we cannot use their static types, and construct
the EditorPlugin object by name.

Special care is taken to make sure that children are removed then added
in that order, both for undo and redo. This is to give the nodes
sequentially-numbered names. Strictly speaking we are relying on the
editor generating the same names when undoing as the nodes had before
they were removed by doing the action; but this is very likely to be the
case, so this is likely to be good enough.