AreaFiller: Instantiate children with GEN_EDIT_STATE_INSTANCE
PackedScene.instantiate() accepts one argument, from the GenEditState
enum. The default value is GEN_EDIT_STATE_DISABLED = 0:

> If passed to instantiate(), blocks edits to the scene state.

In this mode, when the parent scene is saved, the instanced node has a "type"
attribute and a "script" property:

    [node name="Flower" type="Node2D" parent="OnTheGround/Flowerbed" instance=ExtResource("10_cbbm0")]
    position = Vector2(816.928, 202.694)
    script = ExtResource("9_cbbm0")

If, in the editor, you later reopen the scene and save it without changes, these
properties disappear:

    [node name="Flower" parent="OnTheGround/Flowerbed" instance=ExtResource("10_cbbm0")]
    position = Vector2(816.928, 202.694)

The latter serialisation is the one that you get if you instantiate a scene
using the editor UI, rather than in this @tool script.

Another member of the GenEditState enum is GEN_EDIT_STATE_INSTANCE = 1:

> If passed to instantiate(), provides local scene resources to the local scene.

This isn't very well documented but it turns out this causes the scene instanced
in code to be persisted in the second way, as if it had been instanced from the
editor UI.

Pass this flag.