Add custom mouse pointer
This uses an asset included in Tiny Swords. That asset is 64×64 but the
non-transparent area is smaller than 32×32. 32×32 is special on the web
platform:

> Cursor images larger than 32×32 will also only be displayed if the mouse cursor image is entirely located within the page for security reasons.

https://docs.godotengine.org/en/stable/classes/class_input.html#class-input-method-set-custom-mouse-cursor

Add a trimmed copy of the pointer asset, and set this as the pointer.

Doing this in project.godot only overrides the default cursor shape; other
shapes (I-beam, hand, etc.) must be done in code. I believe we don't use any
other shapes in the game so let's use project.godot.