Grappling Hook: Improve handling of loops
Prevent connecting to the immediate previous area. For this, add an
array of excluded areas to the hook control.

Check if the area that is about to be connected has a control that is
connected. In that case, given that there is a single grappling hook in
the game, consider that a loop was created. Pass a new is_loop boolean
parameter to the hooked function of group hook_listeners. This allows
removing the string when a loop is formed, and additionally playing a
blink animation. And also improves the "all hooked" puzzle.