void_runner: Keep zoomed-out camera within level bounds
Interacting with the monk triggers an animation that is intended to zoom
the camera out to be able to see the whole dressed area of the level,
but no further.

The target zoom is a hardcoded multiple of the default zoom, namely 1/4.

However, commit a3040df adjusted the default zoom. As a result, the
effect now zooms out too far and the un-dressed area of the scene can be
seen.

Change the hardcoded magic number.

Fixes https://github.com/endlessm/threadbare/issues/1179