Void runner: Expand dressed area of scene
Previously the dressed area fitted the default aspect ratio, but not far
beyond that ratio in either direction.

Expand the water and the void in all directions. Add one more accessible island
to the left of the starting island (part-covered in void), disconnected from a
second, inaccessible island with the remains of a broken bridge (initially
entirely covered in void).

Extend the camera to the left, top and bottom, but not the right: we don't care
about the water to the right of the monk's island when zoomed out.